refactor(Index): add useTodaysVideo hook
